✨(frontend) display selected threads count in right panel
When threads are selected via checkboxes, the right panel now shows
the selection count with an illustration and a button to disable
selection, instead of showing the thread view or empty placeholder.

The thread selection state has been lifted into a ThreadSelectionProvider
to share it between the thread panel and the right panel pages.
