+++ src/frontend/public/locales/common/fr-FR.json
@@ -44,6 +44,9 @@
   "{{count}} results_one": "{{count}} résultat",
   "{{count}} results_many": "{{count}} résultats",
   "{{count}} results_other": "{{count}} résultats",
+  "{{count}} selected threads_one": "{{count}} conversation sélectionnée",
+  "{{count}} selected threads_many": "{{count}} conversations sélectionnées",
+  "{{count}} selected threads_other": "{{count}} conversations sélectionnées",
   "{{count}} threads have been archived._one": "La conversation a été archivée.",
   "{{count}} threads have been archived._many": "{{count}} conversations ont été archivées.",
   "{{count}} threads have been archived._other": "{{count}} conversations ont été archivées.",
