+++ src/frontend/src/features/blocknote/column-layout-block/_index.scss
@@ -6,7 +6,7 @@
 }
 
 .bn-block-column {
-  --max-width-row: var(--max-block-width, 70ch);
+  --max-width-row: var(--max-block-width, 85ch);
   --min-width-first-column: 60px;
   --min-width-last-column: 200px;
   width: 0;
