+++ src/frontend/src/pages/mailbox/[mailboxId]/thread/[threadId].tsx
@@ -1,9 +1,12 @@
 import { MainLayout } from "@/features/layouts/components/main";
 import { ThreadPanel } from "@/features/layouts/components/thread-panel";
+import { ThreadSelectionPlaceholder } from "@/features/layouts/components/thread-selection-placeholder";
 import { ThreadView } from "@/features/layouts/components/thread-view";
+import { ThreadSelectionProvider, useThreadSelection } from "@/features/providers/thread-selection";
 import { Panel, Group, Separator, useDefaultLayout } from "react-resizable-panels";
 
 const Mailbox = () => {
+    const { selectedThreadIds } = useThreadSelection();
     const { defaultLayout, onLayoutChange } = useDefaultLayout({
         groupId: "threads",
         storage: localStorage,
@@ -16,16 +19,22 @@ const Mailbox = () => {
             </Panel>
             <Separator className="panel__resize-handle" />
             <Panel id="panel-thread-view" className="thread-view-panel">
-                <ThreadView />
+                {selectedThreadIds.size > 0 ? (
+                    <ThreadSelectionPlaceholder />
+                ) : (
+                    <ThreadView />
+                )}
             </Panel>
         </Group>
-    )
-}
+    );
+};
 
 Mailbox.getLayout = function getLayout(page: React.ReactElement) {
     return (
         <MainLayout>
-            {page}
+            <ThreadSelectionProvider>
+                {page}
+            </ThreadSelectionProvider>
         </MainLayout>
     )
 }
